+++ src/pages/course/modal/CourseForm.tsx
@@ -21,6 +21,7 @@ import {
 } from '@/constants/invoice';
 import { useToastStore } from '@/store/feedback/toastStore';
 import ConfirmModal from '@/shared/modal/ConfirmModal';
+import { toDateOnly, toInvoicePaidAt } from './courseFormDateUtils';
 
 interface CourseFormState {
   student: {
@@ -148,15 +149,6 @@ export default function CourseForm({
 
     setIsSubmitting(true);
     try {
-      const formatDateOnly = (value: string) => {
-        const parsed = new Date(value);
-        if (Number.isNaN(parsed.getTime())) return value;
-        const yyyy = parsed.getFullYear();
-        const mm = String(parsed.getMonth() + 1).padStart(2, '0');
-        const dd = String(parsed.getDate()).padStart(2, '0');
-        return `${yyyy}-${mm}-${dd}`;
-      };
-
       const invoiceStatus: 'PAID' | 'UNPAID' =
         form.invoice.status === 'PAID' ? 'PAID' : 'UNPAID';
       const invoiceMethod: 'CARD' | 'CASH' | null =
@@ -164,49 +156,20 @@ export default function CourseForm({
           ? form.invoice.method === 'CARD'
             ? 'CARD'
             : form.invoice.method === 'CASH'
-              ? 'CASH'
-              : null
+            ? 'CASH'
+            : null
           : null;
-      const hasIsoTimezone = (value: string) =>
-        /([zZ]|[+-]\d{2}:\d{2})$/.test(value);
-      const formatTimezoneOffset = (date: Date) => {
-        const offsetMinutes = -date.getTimezoneOffset();
-        const sign = offsetMinutes >= 0 ? '+' : '-';
-        const absMinutes = Math.abs(offsetMinutes);
-        const hours = String(Math.floor(absMinutes / 60)).padStart(2, '0');
-        const minutes = String(absMinutes % 60).padStart(2, '0');
-        return `${sign}${hours}:${minutes}`;
-      };
-      const formatLocalIsoWithOffset = (value: string) => {
-        const parsed = new Date(value);
-        if (Number.isNaN(parsed.getTime())) return value;
-        const yyyy = parsed.getFullYear();
-        const mm = String(parsed.getMonth() + 1).padStart(2, '0');
-        const dd = String(parsed.getDate()).padStart(2, '0');
-        const hh = String(parsed.getHours()).padStart(2, '0');
-        const mi = String(parsed.getMinutes()).padStart(2, '0');
-        const ss = String(parsed.getSeconds()).padStart(2, '0');
-        return `${yyyy}-${mm}-${dd}T${hh}:${mi}:${ss}${formatTimezoneOffset(
-          parsed,
-        )}`;
-      };
-      const invoicePaidAt: string | null =
-        form.invoice.status === 'PAID'
-          ? hasIsoTimezone(form.invoice.paid_at)
-            ? form.invoice.paid_at
-            : form.invoice.paid_at.includes('T')
-              ? formatLocalIsoWithOffset(form.invoice.paid_at)
-              : `${formatDateOnly(form.invoice.paid_at)}T00:00:00${formatTimezoneOffset(
-                  new Date(form.invoice.paid_at),
-                )}`
-          : null;
+      const invoicePaidAt = toInvoicePaidAt(
+        form.invoice.paid_at,
+        invoiceStatus,
+      );
 
       const payload = {
         student_id: form.student.student_id,
         class_type: form.class_type,
         family_discount: form.family_discount,
         lesson_count: form.lesson_count,
-        start_date: formatDateOnly(form.start_date),
+        start_date: toDateOnly(form.start_date),
         schedules: form.schedules,
         invoice: {
           status: invoiceStatus,
